requirements.txt
black==24.10.0
blinker==1.9.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
filelock==3.16.1
flake8==7.1.1
Flask==3.1.0
identify==2.6.5
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
virtualenv==20.28.1
Werkzeug==3.1.3
